Several changes to bring us "current"
A couple of notes - when we're talking about the ML models in the bestpractices section, we'll want to make sure that we make changes to the description of the models, which currently states that we go to tier-3.  This will change eventually, which will necessitate an update to the docs.